Melhorando a API

+++ modulo2/Todo/Controllers/HomeController.cs
@@ -9,51 +9,55 @@ namespace Todo.Controllers
     {
         [HttpGet]
         [Route("/")]
-        public List<TodoModel> Get([FromServices] AppDbContext context)
-        {
-            return context.Todos.ToList();
-        }
+        public IActionResult Get([FromServices] AppDbContext context)
+            => Ok(context.Todos.ToList());
 
         [HttpGet("/{id:int}")]
-        public TodoModel GetById([FromRoute] int id, [FromServices] AppDbContext context)
+        public IActionResult GetById([FromRoute] int id, [FromServices] AppDbContext context)
         {
-            return context.Todos.FirstOrDefault(x => x.Id == id) ?? new TodoModel();
+            var todo = context.Todos.FirstOrDefault(x => x.Id == id);
+            if (todo is null)
+                return NotFound();
+
+            return Ok(todo);
         }
 
         [HttpPost]
         [Route("/")]
-        public TodoModel Post([FromBody] TodoModel model, [FromServices] AppDbContext context)
+        public IActionResult Post([FromBody] TodoModel todo, [FromServices] AppDbContext context)
         {
-            context.Todos.Add(model);
+            context.Todos.Add(todo);
             context.SaveChanges();
-            return model;
+            return Created($"/{todo.Id}", todo);
         }
 
         [HttpPut]
         [Route("/{id:int}")]
-        public TodoModel Put([FromRoute] int id, [FromBody] TodoModel todo, [FromServices] AppDbContext context)
+        public IActionResult Put([FromRoute] int id, [FromBody] TodoModel todo, [FromServices] AppDbContext context)
         {
             var model = context.Todos.FirstOrDefault(x => x.Id == id);
             if (model == null)
-                return todo;
+                return NotFound();
 
             model.Title = todo.Title;
             model.Done = todo.Done;
 
             context.Todos.Update(model);
             context.SaveChanges();
-            return model;
+            return Ok(model);
         }
 
         [HttpDelete]
         [Route("/{id:int}")]
-        public TodoModel Delete([FromRoute] int id, [FromServices] AppDbContext context)
+        public IActionResult Delete([FromRoute] int id, [FromServices] AppDbContext context)
         {
             var model = context.Todos.FirstOrDefault(x => x.Id == id);
+            if (model is null)
+                return NotFound();
 
             context.Todos.Remove(model);
             context.SaveChanges();
-            return model;
+            return Ok(model);
         }
     }
 }
